DDF-4133 Fix backup command
