+++ catalog/core/catalog-core-solrcommands/pom.xml
@@ -26,6 +26,11 @@
     <packaging>bundle</packaging>
     <dependencies>
         <dependency>
+            <groupId>ddf.security.encryption</groupId>
+            <artifactId>security-encryption-api</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
             <groupId>org.apache.httpcomponents</groupId>
             <artifactId>httpclient</artifactId>
             <version>4.5.3</version>
@@ -152,6 +157,12 @@
             <artifactId>log4j-api</artifactId>
             <version>2.11.0</version>
         </dependency>
+        <dependency>
+            <groupId>ddf.platform.solr</groupId>
+            <artifactId>solr-factory-impl</artifactId>
+            <version>${project.version}</version>
+            <scope>compile</scope>
+        </dependency>
     </dependencies>
     <dependencyManagement>
         <dependencies>
@@ -172,7 +183,8 @@
                         <Karaf-Commands>org.codice.ddf.commands.solr</Karaf-Commands>
                         <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
                         <Embed-Dependency>
-                            platform-util
+                            platform-util,
+                            solr-factory-impl
                         </Embed-Dependency>
                         <Private-Package>
                             org.codice.ddf.commands.solr
