+++ platform/solr/solr-factory-impl/src/test/java/org/codice/solr/factory/impl/SolrClientFactoryImplTest.java
@@ -30,6 +30,8 @@ public class SolrClientFactoryImplTest {
 
   @Mock private SolrClient mockClient;
 
+  @Mock private HttpSolrClientFactory mockHttpSolrClientFactory;
+
   // Used to keep and assert on the SolrClientFactory instance newClient would be called on
   private SolrClientFactory solrClientFactory;
 
@@ -41,14 +43,16 @@ public class SolrClientFactoryImplTest {
 
   @Test(expected = IllegalArgumentException.class)
   public void newClientWithNullCoreName() {
-    SolrClientFactoryImpl factory = new SolrClientFactoryImpl(newClientFunction);
+    SolrClientFactoryImpl factory =
+        new SolrClientFactoryImpl(mockHttpSolrClientFactory, newClientFunction);
     factory.newClient(null);
   }
 
   @Test
   public void newEmbeddedSolrClient() {
     System.setProperty("solr.client", "EmbeddedSolrServer");
-    SolrClientFactoryImpl factory = new SolrClientFactoryImpl(newClientFunction);
+    SolrClientFactoryImpl factory =
+        new SolrClientFactoryImpl(mockHttpSolrClientFactory, newClientFunction);
 
     SolrClient client = factory.newClient("core");
     assertThat(solrClientFactory, is(instanceOf(EmbeddedSolrFactory.class)));
@@ -58,17 +62,19 @@ public class SolrClientFactoryImplTest {
   @Test
   public void newHttpSolrClient() {
     System.setProperty("solr.client", "HttpSolrClient");
-    SolrClientFactoryImpl factory = new SolrClientFactoryImpl(newClientFunction);
+    SolrClientFactoryImpl factory =
+        new SolrClientFactoryImpl(mockHttpSolrClientFactory, newClientFunction);
 
     SolrClient client = factory.newClient("core");
-    assertThat(solrClientFactory, is(instanceOf(HttpSolrClientFactory.class)));
+    assertThat(solrClientFactory, is(mockHttpSolrClientFactory));
     assertThat(client, is(mockClient));
   }
 
   @Test
   public void newCloudSolrClient() {
     System.setProperty("solr.client", "CloudSolrClient");
-    SolrClientFactoryImpl factory = new SolrClientFactoryImpl(newClientFunction);
+    SolrClientFactoryImpl factory =
+        new SolrClientFactoryImpl(mockHttpSolrClientFactory, newClientFunction);
 
     SolrClient client = factory.newClient("core");
     assertThat(solrClientFactory, is(instanceOf(SolrCloudClientFactory.class)));
@@ -78,10 +84,11 @@ public class SolrClientFactoryImplTest {
   @Test
   public void newClientWithUnknownClientType() {
     System.setProperty("solr.client", "Unknown");
-    SolrClientFactoryImpl factory = new SolrClientFactoryImpl(newClientFunction);
+    SolrClientFactoryImpl factory =
+        new SolrClientFactoryImpl(mockHttpSolrClientFactory, newClientFunction);
 
     SolrClient client = factory.newClient("core");
-    assertThat(solrClientFactory, is(instanceOf(HttpSolrClientFactory.class)));
+    assertThat(solrClientFactory, is(mockHttpSolrClientFactory));
     assertThat(client, is(mockClient));
   }
 }
